+++ tests/test_coordinator_coverage.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from datetime import UTC, datetime
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
@@ -89,44 +88,6 @@ async def test_get_client_method_from_client():
 # ---------------------------------------------------------------------------
 
 
-def test_compute_register_groups_safe_scan():
-    """safe_scan=True produces per-register (addr, length) tuples (lines 1026-1047)."""
-    coord = _make_coordinator(safe_scan=True)
-    coord.available_registers = {
-        "input_registers": {"outside_temperature"},
-        "holding_registers": set(),
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-    coord._compute_register_groups()
-    groups = coord._register_groups.get("input_registers", [])
-    assert isinstance(groups, list)
-    assert len(groups) >= 1
-    addr, length = groups[0]
-    assert isinstance(addr, int)
-    assert isinstance(length, int)
-
-
-def test_compute_register_groups_safe_scan_unknown_register():
-    """Unknown register in safe_scan mode is skipped gracefully."""
-    coord = _make_coordinator(safe_scan=True)
-    coord.available_registers = {
-        "input_registers": {"__unknown_reg_xyz__"},
-        "holding_registers": set(),
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-    coord._compute_register_groups()
-    # No exception should be raised; groups for input_registers is empty since reg not in map
-    groups = coord._register_groups.get("input_registers", [])
-    assert isinstance(groups, list)
-
-
-# ---------------------------------------------------------------------------
-# Group I — _test_connection exception handlers (lines 1125-1141)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_test_connection_modbus_io_cancelled_skips():
     """ModbusIOException with 'cancelled' message is swallowed (lines 1125-1130)."""
@@ -179,128 +140,6 @@ async def test_test_connection_modbus_io_non_cancelled_raises():
 
 # ---------------------------------------------------------------------------
 # Group N — calculate_power_consumption (lines 1856-1881)
-# ---------------------------------------------------------------------------
-
-
-def test_calculate_power_consumption_basic():
-    """calculate_power_consumption returns float when dac_supply/exhaust provided."""
-    coord = _make_coordinator()
-    data = {"dac_supply": 5.0, "dac_exhaust": 5.0}
-    result = coord.calculate_power_consumption(data)
-    assert result is not None
-    assert isinstance(result, float)
-    assert result > 0
-
-
-def test_calculate_power_consumption_with_heater_and_cooler():
-    """Heater and cooler voltages contribute to power (lines 1876-1879)."""
-    coord = _make_coordinator()
-    data = {"dac_supply": 8.0, "dac_exhaust": 7.0, "dac_heater": 5.0, "dac_cooler": 3.0}
-    result = coord.calculate_power_consumption(data)
-    assert result is not None
-    assert result > 0
-
-
-def test_calculate_power_consumption_missing_keys_returns_none():
-    """KeyError on missing dac_supply/exhaust returns None (lines 1861-1862)."""
-    coord = _make_coordinator()
-    result = coord.calculate_power_consumption({})
-    assert result is None
-
-
-def test_calculate_power_consumption_invalid_type_returns_none():
-    """TypeError on non-numeric values returns None."""
-    coord = _make_coordinator()
-    result = coord.calculate_power_consumption({"dac_supply": "bad", "dac_exhaust": 5.0})
-    assert result is None
-
-
-# ---------------------------------------------------------------------------
-# Group O — _post_process_data branches (lines 1883-1925)
-# ---------------------------------------------------------------------------
-
-
-def test_post_process_data_zero_division_error():
-    """ZeroDivisionError when exhaust == outside is caught silently (lines 1894-1898)."""
-    coord = _make_coordinator()
-    data = {
-        "outside_temperature": 20.0,
-        "supply_temperature": 22.0,
-        "exhaust_temperature": 20.0,  # Same as outside → ZeroDivisionError
-    }
-    result = coord._post_process_data(data)
-    # Should not raise; calculated_efficiency is absent
-    assert "calculated_efficiency" not in result
-
-
-def test_post_process_data_efficiency_calculated():
-    """Heat recovery efficiency is calculated when temperatures differ."""
-    coord = _make_coordinator()
-    data = {
-        "outside_temperature": 0.0,
-        "supply_temperature": 18.0,
-        "exhaust_temperature": 20.0,
-    }
-    result = coord._post_process_data(data)
-    assert "calculated_efficiency" in result
-    assert 0 <= result["calculated_efficiency"] <= 100
-
-
-def test_post_process_data_flow_balance():
-    """Flow balance is calculated from supply and exhaust flow rates."""
-    coord = _make_coordinator()
-    data = {"supply_flow_rate": 100, "exhaust_flow_rate": 75}
-    result = coord._post_process_data(data)
-    assert result["flow_balance"] == 25
-    assert result["flow_balance_status"] == "supply_dominant"
-
-
-def test_post_process_data_flow_balance_exhaust_dominant():
-    """Flow balance status is exhaust_dominant when exhaust > supply."""
-    coord = _make_coordinator()
-    data = {"supply_flow_rate": 80, "exhaust_flow_rate": 100}
-    result = coord._post_process_data(data)
-    assert result["flow_balance_status"] == "exhaust_dominant"
-
-
-def test_post_process_data_flow_balance_balanced():
-    """Flow balance status is balanced when diff < 10."""
-    coord = _make_coordinator()
-    data = {"supply_flow_rate": 100, "exhaust_flow_rate": 98}
-    result = coord._post_process_data(data)
-    assert result["flow_balance_status"] == "balanced"
-
-
-def test_post_process_data_flow_balance_string_values():
-    """flow_balance must not crash when register returns a string value."""
-    coord = _make_coordinator()
-    data = {"supply_flow_rate": "invalid", "exhaust_flow_rate": 75}
-    result = coord._post_process_data(data)
-    # Should silently skip — no crash, no flow_balance key
-    assert "flow_balance" not in result
-
-
-def test_post_process_data_power_calculation():
-    """Power is estimated and energy accumulated when DAC values provided."""
-    coord = _make_coordinator()
-    data = {"dac_supply": 5.0, "dac_exhaust": 5.0}
-    result = coord._post_process_data(data)
-    assert "estimated_power" in result
-    assert "total_energy" in result
-
-
-def test_post_process_data_timezone_aware_timestamp():
-    """Timezone-aware last timestamp is handled correctly (lines 1916-1919)."""
-    coord = _make_coordinator()
-    # Set a timezone-aware last timestamp
-    coord._last_power_timestamp = datetime(2024, 1, 1, 12, 0, 0, tzinfo=UTC)
-    data = {"dac_supply": 3.0, "dac_exhaust": 3.0}
-    result = coord._post_process_data(data)
-    assert "estimated_power" in result
-
-
-# ---------------------------------------------------------------------------
-# Group R — async_write_temporary_* (lines 2320-2366)
 # ---------------------------------------------------------------------------
 
 
@@ -407,38 +246,6 @@ async def test_test_connection_modbus_exception_raises():
 
 # ---------------------------------------------------------------------------
 # _apply_scan_cache exception branch (lines 985-986)
-# ---------------------------------------------------------------------------
-
-
-def test_apply_scan_cache_normalise_exception():
-    """TypeError in _normalise_available_registers returns False (lines 985-986)."""
-    coord = _make_coordinator()
-    with patch.object(
-        coord,
-        "_normalise_available_registers",
-        side_effect=TypeError("bad"),
-    ):
-        result = coord._apply_scan_cache({"available_registers": {"input_registers": ["mode"]}})
-    assert result is False
-
-
-def test_apply_scan_cache_invalid_capabilities():
-    """Invalid capabilities dict in cache is caught silently (lines 994-995)."""
-    from custom_components.thessla_green_modbus.scanner import DeviceCapabilities
-
-    coord = _make_coordinator()
-    with patch.object(DeviceCapabilities, "__init__", side_effect=TypeError("bad")):
-        result = coord._apply_scan_cache(
-            {
-                "available_registers": {"input_registers": ["mode"]},
-                "capabilities": {"bad_key": 1},
-            }
-        )
-    assert result is True  # overall still succeeds
-
-
-# ---------------------------------------------------------------------------
-# Pass 15 — Sekcja A: __init__ jitter else branch (lines 326-329)
 # ---------------------------------------------------------------------------
 
 
@@ -710,42 +517,6 @@ async def test_async_write_register_error_response_retries():
     assert transport.write_register.call_count == 2  # retried once
 
 
-def test_post_process_data_type_error_in_efficiency():
-    """TypeError in efficiency calculation is caught (lines 1897-1898)."""
-    coord = _make_coordinator()
-    data = {
-        "outside_temperature": "not_a_number",  # triggers TypeError
-        "supply_temperature": 20.0,
-        "exhaust_temperature": 25.0,
-    }
-    result = coord._post_process_data(data)
-    assert "calculated_efficiency" not in result
-
-
-def test_post_process_data_non_datetime_last_timestamp():
-    """Non-datetime _last_power_timestamp → elapsed=0.0 (line 1914)."""
-    coord = _make_coordinator()
-    coord._last_power_timestamp = "not_a_datetime"
-    data = {"dac_supply": 3.0, "dac_exhaust": 3.0}
-    result = coord._post_process_data(data)
-    assert "estimated_power" in result
-    assert "total_energy" in result
-
-
-def test_post_process_data_naive_now_aware_last_ts():
-    """Naive _utcnow with aware last_ts → adds UTC tz to now (line 1919)."""
-    coord = _make_coordinator()
-    coord._last_power_timestamp = datetime(2024, 1, 1, 12, 0, 0, tzinfo=UTC)
-    # Patch _utcnow to return naive datetime
-    with patch(
-        "custom_components.thessla_green_modbus.coordinator.coordinator._utcnow",
-        return_value=datetime(2024, 1, 1, 12, 0, 30),  # naive
-    ):
-        data = {"dac_supply": 3.0, "dac_exhaust": 3.0}
-        result = coord._post_process_data(data)
-    assert "estimated_power" in result
-
-
 def test_create_consecutive_groups_empty():
     """Empty registers dict returns [] (line 1931)."""
     from custom_components.thessla_green_modbus._coordinator_register_processing import (
@@ -775,29 +546,6 @@ async def test_call_modbus_no_client_raises():
 
     with pytest.raises(ConnectionException):
         await coord._call_modbus(dummy, 100, count=1)
-
-
-def test_get_device_info_model_from_entry():
-    """get_device_info uses entry.options when device_info has no model (line 2539)."""
-    coord = _make_coordinator()
-    from unittest.mock import MagicMock
-
-    entry = MagicMock()
-    entry.options = {"model": "Thessla Air 350"}
-    entry.data = {}
-    coord.entry = entry
-    coord.device_scan_result = {}
-    coord.device_info = {}
-    info = coord.get_device_info()
-    assert info["model"] == "Thessla Air 350"
-
-
-def test_compat_device_info_getattr_key_error():
-    """_CompatDeviceInfo.__getattr__ raises AttributeError for missing key (lines 2552-2555)."""
-    coord = _make_coordinator()
-    info = coord.get_device_info()
-    with pytest.raises(AttributeError):
-        _ = info.nonexistent_attribute_xyz
 
 
 @pytest.mark.asyncio
@@ -846,46 +594,6 @@ async def test_read_discrete_inputs_transport_returns_result():
 
 # ---------------------------------------------------------------------------
 # Pass 16 — B2: _normalise_available_registers invalid type (line 968)
-# ---------------------------------------------------------------------------
-
-
-def test_compute_register_groups_safe_scan_key_error():
-    """KeyError in get_register_definition with safe_scan=True (lines 1035-1037)."""
-    coord = _make_coordinator(safe_scan=True)
-    coord.available_registers = {"holding_registers": {"mode"}}
-    coord._register_maps = {"holding_registers": {"mode": 100}}
-    with patch(
-        "custom_components.thessla_green_modbus.coordinator.get_register_definition",
-        side_effect=KeyError("mode"),
-    ):
-        coord._compute_register_groups()
-    assert "holding_registers" in coord._register_groups
-
-
-def test_compute_register_groups_non_safe_addr_none():
-    """addr=None in register map hits continue (line 1053)."""
-    coord = _make_coordinator(safe_scan=False)
-    coord.available_registers = {"holding_registers": {"unknown_reg"}}
-    coord._register_maps = {"holding_registers": {}}  # addr will be None
-    coord._compute_register_groups()
-    assert coord._register_groups.get("holding_registers", []) == []
-
-
-def test_compute_register_groups_non_safe_key_error():
-    """KeyError in get_register_definition with safe_scan=False (lines 1057-1059)."""
-    coord = _make_coordinator(safe_scan=False)
-    coord.available_registers = {"holding_registers": {"mode"}}
-    coord._register_maps = {"holding_registers": {"mode": 100}}
-    with patch(
-        "custom_components.thessla_green_modbus.coordinator.get_register_definition",
-        side_effect=KeyError("mode"),
-    ):
-        coord._compute_register_groups()
-    assert "holding_registers" in coord._register_groups
-
-
-# ---------------------------------------------------------------------------
-# Pass 16 — B4: _test_connection paths (lines 1111, 1122)
 # ---------------------------------------------------------------------------
